增加了实时显示分数的功能，将时间分数都化为了整数，对部分UI做了点没有大碍的微调，删掉了武器的过热和散热功能，对敌人的刷新数量和刷新速度做了调整，对爆炸震动的幅度做了调整并且增加了个随机变量进行控制。

+++ Assets/Script/Global/GameManager.cs
@@ -54,7 +54,7 @@ public class GameManager : SingleBehaviour<GameManager>
 	public static void Gameover()
 	{
 		var state = Instance.state;
-		state.LastPlayTime = GameTime.PassedTime;
+		state.LastPlayTime = (int)GameTime.PassedTime;
 
 		if (state.LongestPlayTime < state.LastPlayTime)
 			state.LongestPlayTime = state.LastPlayTime;
